Final Commit
My code works as intended.  It runs the game snake on pygame. The snake grows as you eat food and continues to grow until you run into your own tail or run into the wall. once the game ends you have a choice to play again or quit the game.